trying to fix up tests

+++ Cpu.cpp
@@ -142,10 +142,10 @@ void Cpu::execute(const instruction_union& instr)
 	{
 		case cpu_instructions::ADDI:
 		{
-			unsigned short immediate = instr.immediate_instruction.immediate;
+			unsigned int immediate = (short)instr.immediate_instruction.immediate;
 			int rs_value = register_file.get_register(instr.immediate_instruction.rs);
 
-			unsigned int value = rs_value + (short)immediate;
+			unsigned int value = rs_value + immediate;
 
 			// check for overflow
 			{
@@ -173,10 +173,10 @@ void Cpu::execute(const instruction_union& instr)
 
 		case cpu_instructions::ADDIU:
 		{
-			unsigned short immediate = instr.immediate_instruction.immediate;
+			unsigned int immediate = (short)instr.immediate_instruction.immediate;
 			unsigned int rs_value = register_file.get_register(instr.immediate_instruction.rs);
 
-			unsigned int value = rs_value + (short)immediate;
+			unsigned int value = rs_value + immediate;
 
 			register_file.set_register(instr.immediate_instruction.rt, value);
 		} break;
